refactor: 자동차 길찾기 API, 다중 길찾기 API 호출 클라이언트를 클래스로 분리
